Pull this pleaseee

+++ abstract-crawler/inc/AbstractCrawler.h
@@ -42,9 +42,7 @@ public:
 
     static size_t count_tags(const char* html, size_t html_size);
 
-    void add_from_file(const std::string& filename, int16_t num_websites);
-
-
+    void add_from_file(const std::string& filename, int64_t num_websites);
 };
